Add error handling

+++ app/services/file_manager/zip_creator.rb
@@ -9,12 +9,11 @@ module FileManager
     end
 
     def call
+      raise FileNotFoundError, "File not found at path: #{file_path}" unless File.exist?(file_path)
+
       file_name = File.basename(file_path)
       zip_name = File.basename(file_name, File.extname(file_name))
       zip_file_path = Rails.root.join("tmp", "#{zip_name}.zip")
-
-      # Override the existing zip file if it already exists
-      File.delete(zip_file_path) if File.exist?(zip_file_path)
 
       Zip::File.open(zip_file_path, Zip::File::CREATE) do |zipfile|
         zipfile.add(file_name, file_path)
